v2/app/requirements.txt
Flask==1.1.1
Flask-Session==0.3.1
healthcheck==1.3.3
psycopg2-binary==2.8.4
redis==3.3.11
six==1.13.0
